MEMORY CARD GAME
Got the idea to create this game on the plane ride to Pittsburgh.  My son was getting bored of the tablet on the plane ride, so I wanted to see how quickly I could create the game.  All I had was my phone and bluetooth keyboard LOL.

I finished it before the plane landed, but didn't have the json file or asset folder with the images to reference, so there were just a bunch of blocks on the screen. **sad face**

As soon as I got home, I built it and downloaded the assets to play with my son. After a couple rounds, he asked if he could have 'ghost pikachu' on there, so I downloaded some pictures of ghost pokemon and added in the pokemonTCG backing to be the back of the card.

After a few rounds of that (which he LOVED), he asked if a BIG ghost pikachu could popup.  I created a "You Win" alert when the score reaches 9, but I want to add something like a lightbox that pops up with the ghost pikachu and a YOU WIN banner.  I think he'd like that :D

--created following tutorial; edited additional featuers on my own :D
https://dev.to/javascriptacademy/creating-a-memory-card-game-with-html-css-and-javascript-57g1